reuse create project code
